moved the checkboxes into form folder and placed form files into TextImput folder
