+++ client/src/components/Forms/Checkboxes/Checkbox.jsx
@@ -3,7 +3,7 @@ import { useId } from "react";
 
 export default function Checkbox({ checkboxLabel, id, name }) {
   const inputId = id ?? useId();
-  const inputName = name ?? id;
+  const inputName = name ?? inputId;
   return (
     <div className={styles.checkbox}>
       <input type="checkbox" id={inputId} name={inputName} />
