--- client/src/components/Forms/Form
@@ -1,12 +1,0 @@
-import styles from "./Form.module.css";
-
-export default function Form({ formLable }) {
-  return (
-  <>
-    <form>
-        <label for={ formLable }>{ formLable }</label>
-            <input type="text" id={ formLable } name={ formLable }/>
-    </form>;
-  </>
-  )
-}
