--- client/src/components/Forms/Form.module.css
@@ -1,6 +1,0 @@
-.form{
-    display: flex;
-    flex-direction: column;
-    text-align: left;
-    width: 10rem;
-}
